+++ barretenberg/cpp/src/barretenberg/relations/ecc_vm/ecc_wnaf_relation.cpp
