+++ barretenberg/cpp/src/barretenberg/relations/lookup_relation.hpp
@@ -1,5 +1,5 @@
 #pragma once
-#include "barretenberg/proof_system/relations/relation_types.hpp"
+#include "barretenberg/relations/relation_types.hpp"
 
 namespace proof_system {
