+++ barretenberg/cpp/src/barretenberg/relations/nested_containers.test.cpp
@@ -1,4 +1,4 @@
-#include "barretenberg/proof_system/relations/nested_containers.hpp"
+#include "barretenberg/relations/nested_containers.hpp"
 #include "barretenberg/ecc/curves/bn254/fr.hpp"
 #include "barretenberg/polynomials/univariate.hpp"
 #include <gtest/gtest.h>
