+++ barretenberg/cpp/src/barretenberg/relations/translator_vm/goblin_translator_relation_consistency.test.cpp
@@ -12,12 +12,12 @@
  *
  */
 #include "barretenberg/ecc/curves/bn254/fr.hpp"
-#include "barretenberg/proof_system/relations/relation_parameters.hpp"
-#include "barretenberg/proof_system/relations/translator_vm/translator_decomposition_relation.hpp"
-#include "barretenberg/proof_system/relations/translator_vm/translator_extra_relations.hpp"
-#include "barretenberg/proof_system/relations/translator_vm/translator_gen_perm_sort_relation.hpp"
-#include "barretenberg/proof_system/relations/translator_vm/translator_non_native_field_relation.hpp"
-#include "barretenberg/proof_system/relations/translator_vm/translator_permutation_relation.hpp"
+#include "barretenberg/relations/relation_parameters.hpp"
+#include "barretenberg/relations/translator_vm/translator_decomposition_relation.hpp"
+#include "barretenberg/relations/translator_vm/translator_extra_relations.hpp"
+#include "barretenberg/relations/translator_vm/translator_gen_perm_sort_relation.hpp"
+#include "barretenberg/relations/translator_vm/translator_non_native_field_relation.hpp"
+#include "barretenberg/relations/translator_vm/translator_permutation_relation.hpp"
 #include <gtest/gtest.h>
 
 using namespace proof_system;
